requirements.txt
pytesseract~=0.3.10
numpy~=1.26.4
opencv-python~=4.9.0.80
requests~=2.31.0
pyyaml~=6.0.1
discord.py~=2.3.2